+++ Magnetometer/unitTest.py
