+++ Magnetorquer/magnetorquer.py
