--- mpu9250/__init__.py
@@ -1,3 +1,0 @@
-__version__ = '0.0.1'
-
-from mpu9250 import mpu9250
