--- portClear.py
@@ -1,6 +1,0 @@
-import config
-import socket
-
-soc = socket.socket()
-soc.bind(("",config.HTTP_PORT))
-soc.close()
