--- supplier.py
@@ -1,4 +1,0 @@
-def supply(output):
-	nothing  = 0
-	#implement supplier control code here:
-
